move into separate build_summary file
